use float for top_k per semconvs https://github.com/open-telemetry/semantic-conventions-genai/blob/main/docs/registry/attributes/gen-ai.md#gen-ai-request-top-k

+++ util/opentelemetry-util-genai/src/opentelemetry/util/genai/_retrieval_invocation.py
@@ -73,7 +73,7 @@ class RetrievalInvocation(GenAIInvocation):
         self.request_model = request_model
         self.server_address = server_address
         self.server_port = server_port
-        self.top_k: int | None = None
+        self.top_k: float | None = None
         self.query_text: str | None = None
         self.documents: Any = None
         self._start(self._get_base_attributes())
